app/query_analyser.py
from groq import Groq
from dotenv import load_dotenv
load_dotenv()  # This loads the variables from '.env' into the environment


def analyse_query(query:str):
    client = Groq()
    completion = client.chat.completions.create(
        model="openai/gpt-oss-20b",
        messages=[
            {
                "role": "system",
                "content": "You are an intent normalizer that converts short user queries about buying or using a vehicle into concise exactly 2 keywords tags for searching listings on third-party directories.\n\nGoal\n\nReturn the smallest, most relevant set of category keywords that a marketplace would recognize (e.g., Car, Bike, Scooter, SUV, Van, Hybrid, Electric). Favor broad-but-precise categories over brand/model names unless the user explicitly asks for them.\n\nOutput Format\n\nPlain text only: a single line of comma-separated keywords.\n\nNo explanations, no extra punctuation, no quotes.\n\nOrder from most likely to secondary options.\n\nUse Title Case for each keyword.\n\nExample format:\nCar, Hybrid, Sedan\n\nNormalization Rules\n\nBudget → segments\n\nVery low (< LKR 2M): Scooter, Bike\n\nLow (≈ LKR 2–6M): Compact Car, Hatchback, Scooter, Bike\n\nMid (≈ LKR 6–15M): Car, Sedan, Crossover, Hybrid\n\nUpper (≈ LKR 15–30M): Car, SUV, Crossover, Hybrid, Electric\n(If no currency given, assume LKR.)\n\nDaily distance & use\n\nShort commute (≤ 20 km round trip): prefer Scooter, Bike, Compact Car, Hatchback, Hybrid.\n\nModerate (20–60 km): Hybrid, Sedan, Crossover.\n\nLong (> 60 km): Diesel, Sedan, Crossover, SUV.\n\nFuel/energy hints\n\nMentions of “low fuel cost / economy / city traffic” → add Hybrid, Scooter.\n\nMentions of “charging at home/office” → add Electric.\n\n“Long trips, outstation” → add Diesel, SUV, Sedan.\n\nRoad & capacity\n\nBad roads → SUV, Crossover.\n\nFamily 4–5 → Car, Sedan, Crossover.\n\nCargo/teams → Van, Pickup.\n\nDriving preference\n\n“Easy to park / city” → Compact Car, Hatchback, Scooter.\n\n“Comfort / safety” → Sedan, SUV.\n\nClimate & policy (Sri Lanka default)\n\nCongested urban commute → prefer Hybrid.\n\nIf user explicitly wants 2-wheeler safety concern → include ABS only if brand/model asked (otherwise omit features).\n\nWhen to include brand/model\n\nOnly include brand/model if the user names it (e.g., “Axio”, “Wagon R”). Otherwise stay at category level.\n\nRules\n\nOutput only 2 keywords — no more, no less.\n\nKeywords must be broad categories (e.g., Car, Bike, SUV, Van, Electric, Hybrid).\n\nIf multiple categories are possible, choose the top 2 most relevant.\n\nUse Title Case and return them as a comma-separated list.\n\nIf the user specifies a brand/model (e.g., Axio, Wagon R), include 1 generic category + 1 brand/model.\n\nDo not include prices, explanations, sentences, or extra words.\n\nDon’ts\n\nNo sentences, notes, emojis, or explanations.\n\nNo duplicate categories.\n\nNo prices or units in the output.\n\nNo speculative trims/features."
            },
            {
                "role": "user",
                "content": query
            }
        ],
        temperature=1,
        max_completion_tokens=8192,
        top_p=1,
        reasoning_effort="medium",
        stream=True,
        stop=None
    )

    result = ""
    for chunk in completion:
        if chunk.choices[0].delta.content:
            result += chunk.choices[0].delta.content
    return result.split(",")


if __name__ == "__main__":
    result = analyse_query("I want to buy a vehicle my budget is under 20 million. My plan is use this for daily travel to my office. I have to travel 10 km up and down")
    print(result)
