requirements.txt
requests
bs4
fastapi[standard]
pydantic
starlette
groq
python-dotenv